+++ packages/thumbprint-react/components/UIAction/get-anchor-props.ts
@@ -1,18 +1,34 @@
+import split from 'lodash/split';
 import isThumbtackUrl from './is-thumbtack-url';
 
-const getRel = (url?: string, target?: string, shouldOpenInNewTab = false): string | undefined => {
+const getRel = (
+    url?: string,
+    target?: string,
+    shouldOpenInNewTab = false,
+    rel?: string,
+): string | undefined => {
+    // Use a set to prevent duplicate rel values from being defined
+    let relSet = new Set();
+    if (rel) {
+        // handle possibility of multiple rel values being provided
+        const relStrings = split(rel, ' ');
+        relSet = new Set(relStrings);
+    }
     if (shouldOpenInNewTab || target === '_blank') {
-        if (isThumbtackUrl(url)) {
-            // There are performance benefits of adding `rel="noopener"`.
-            // https://jakearchibald.com/2016/performance-benefits-of-rel-noopener/
-            return 'noopener';
-        }
-
+        // There are performance benefits of adding `rel="noopener"`.
+        // https://jakearchibald.com/2016/performance-benefits-of-rel-noopener/
         // https://mathiasbynens.github.io/rel-noopener/
         // https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/jsx-no-target-blank.md
-        return 'noopener noreferrer';
+        relSet.add('noopener');
+
+        if (!isThumbtackUrl(url)) {
+            relSet.add('noreferrer');
+        }
     }
 
+    if (relSet.size > 0) {
+        return [...relSet].join(' ');
+    }
     return undefined;
 };
 
@@ -31,20 +47,22 @@ const getAnchorProps = ({
     shouldOpenInNewTab,
     to,
     onClick,
+    rel,
     target,
 }: {
     isDisabled?: boolean;
     shouldOpenInNewTab?: boolean;
     to?: string;
     onClick?: (event: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => void;
+    rel?: string;
     target?: string;
 }): AnchorProps => ({
     onClick,
     href: isDisabled ? undefined : to,
-    target: target || (shouldOpenInNewTab ? '_blank' : '_self'),
     // NOTE use `noopener` even for internal links in new tabs to prevent potential slowdown of
     // new tab https://mathiasbynens.github.io/rel-noopener/
-    rel: getRel(to, target, shouldOpenInNewTab),
+    rel: getRel(to, target, shouldOpenInNewTab, rel),
+    target: target || (shouldOpenInNewTab ? '_blank' : '_self'),
 });
 
 export default getAnchorProps;