+++ packages/thumbprint-react/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## Unreleased
 
+## 14.6.0 - 2022-03-21
+
+### Changed
+
+-   [Patch] Refactor how default `rel` attribute values are defined
+
+### Added
+
+-   [Minor] Add optional `rel` prop to `Link` and `ThemedLink` components.
+
+## 14.5.0 - 2022-03-17
+
+### Changed
+
+-   [Minor] Rename all component prop interfaces and export them. This is a minor change since the interfaces were private before, so renaming them will not break any client code.
+-   [Patch] Update Type definitions for `Dropdown` to support generic types.
+
 ## 14.4.1 - 2021-09-03
 
 ### Changed